requirements.txt
python-dotenv
flask
openai==1.35.15
requests
streamlit
beautifulsoup4
groq #alternative to openai, same modules as openai lib

